refactor(api): move get learner csv template into a bounded context
