+++ api/tests/prescription/learner-management/unit/application/sup-organization-management-controller_test.js
@@ -1,6 +1,7 @@
 import { supOrganizationManagementController } from '../../../../../src/prescription/learner-management/application/sup-organization-management-controller.js';
 import { usecases } from '../../../../../src/prescription/learner-management/domain/usecases/index.js';
 import { catchErr, expect, hFake, sinon } from '../../../../test-helper.js';
+import { getI18n } from '../../../../tooling/i18n/i18n.js';
 
 describe('Unit | Controller | sup-organization-management-controller', function () {
   let organizationId;
@@ -224,6 +225,49 @@ describe('Unit | Controller | sup-organization-management-controller', function 
       expect(response.statusCode).to.be.equal(200);
 
       expect(logErrorWithCorrelationIdsStub).to.have.been.calledWith(error);
+    });
+  });
+  describe('#getOrganizationLearnersCsvTemplate', function () {
+    const userId = 1;
+    const organizationId = 2;
+    const request = {
+      query: {
+        accessToken: 'token',
+      },
+      params: {
+        id: organizationId,
+      },
+    };
+    let dependencies;
+
+    beforeEach(function () {
+      sinon.stub(usecases, 'getOrganizationLearnersCsvTemplate').resolves('template');
+
+      const tokenServiceStub = {
+        extractUserId: sinon.stub(),
+      };
+      tokenServiceStub.extractUserId.returns(userId);
+
+      dependencies = {
+        tokenService: tokenServiceStub,
+      };
+    });
+
+    it('should return a response with correct headers', async function () {
+      // when
+      request.i18n = getI18n();
+      hFake.request = {
+        path: '/api/organizations/2/sup-organization-learners/csv-template',
+      };
+      const response = await supOrganizationManagementController.getOrganizationLearnersCsvTemplate(
+        request,
+        hFake,
+        dependencies,
+      );
+
+      // then
+      expect(response.headers['Content-Type']).to.equal('text/csv;charset=utf-8');
+      expect(response.headers['Content-Disposition']).to.equal('attachment; filename=modele-import.csv');
     });
   });
 });
