+++ api/tests/prescription/learner-management/unit/domain/usecases/get-organization-learners-csv-template_test.js
@@ -1,9 +1,9 @@
 import _ from 'lodash';
 
-import { UserNotAuthorizedToAccessEntityError } from '../../../../lib/domain/errors.js';
-import { getOrganizationLearnersCsvTemplate } from '../../../../lib/domain/usecases/get-organization-learners-csv-template.js';
-import { catchErr, domainBuilder, expect, sinon } from '../../../test-helper.js';
-import { getI18n } from '../../../tooling/i18n/i18n.js';
+import { UserNotAuthorizedToAccessEntityError } from '../../../../../../lib/domain/errors.js';
+import { getOrganizationLearnersCsvTemplate } from '../../../../../../src/prescription/learner-management/domain/usecases/get-organization-learners-csv-template.js';
+import { catchErr, domainBuilder, expect, sinon } from '../../../../../test-helper.js';
+import { getI18n } from '../../../../../tooling/i18n/i18n.js';
 
 describe('Unit | UseCase | get-organization-learners-csv-template', function () {
   let userId, organizationId, membershipRepository, i18n;
