+++ api/tests/unit/application/organizations/index_test.js
@@ -673,70 +673,6 @@ describe('Unit | Router | organization-router', function () {
     });
   });
 
-  describe('GET /api/organizations/{id}/sup-organization-learners/csv-template', function () {
-    let httpTestServer;
-
-    beforeEach(async function () {
-      sinon
-        .stub(organizationController, 'getOrganizationLearnersCsvTemplate')
-        .callsFake((request, h) => h.response('ok').code(200));
-
-      httpTestServer = new HttpTestServer();
-      await httpTestServer.register(moduleUnderTest);
-    });
-
-    it('should call the organization controller to csv template', async function () {
-      // given
-      const method = 'GET';
-      const url = '/api/organizations/1/sup-organization-learners/csv-template?accessToken=token';
-
-      // when
-      const response = await httpTestServer.request(method, url);
-
-      // then
-      expect(response.statusCode).to.equal(200);
-      expect(organizationController.getOrganizationLearnersCsvTemplate).to.have.been.calledOnce;
-    });
-
-    describe('When parameters are not valid', function () {
-      it('should throw an error when id is not a number', async function () {
-        // given
-        const method = 'GET';
-        const url = '/api/organizations/ABC/sup-organization-learners/csv-template?accessToken=token';
-
-        // when
-        const response = await httpTestServer.request(method, url);
-
-        // then
-        expect(response.statusCode).to.equal(400);
-      });
-
-      it('should throw an error when id is null', async function () {
-        // given
-        const method = 'GET';
-        const url = '/api/organizations/null/sup-organization-learners/csv-template?accessToken=token';
-
-        // when
-        const response = await httpTestServer.request(method, url);
-
-        // then
-        expect(response.statusCode).to.equal(400);
-      });
-
-      it('should throw an error when access token is not specified', async function () {
-        // given
-        const method = 'GET';
-        const url = '/api/organizations/1/sup-organization-learners/csv-template';
-
-        // when
-        const response = await httpTestServer.request(method, url);
-
-        // then
-        expect(response.statusCode).to.equal(400);
-      });
-    });
-  });
-
   describe('DELETE /api/organizations/{id}/invitations/{invitationId}', function () {
     it('should call the cancel organization invitation controller', async function () {
       // given
